examples/healthcare_voice_agent/requirements.txt
langchain==0.3.26
langgraph==0.4.9
langchain-openai==0.3.25
openai==1.91.0
python-dotenv==1.1.1
speechrecognition[whisper-local]==3.14.3
coqui-tts==0.26.2
numpy==2.0.0
soundfile==0.12.1 
traceroot==0.0.3a6